+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 
 - Enum-Unterstützung wurde nach gdv-xport-deprecated verschoben
 - eine Reihe Methoden, die für v6 als "deprecated" gekennzeichnet waren, wurden entfernt
+- Property "gdv.numfeld.fill-blanks" wurde wieder entfernt
 
 
 ## [5.4.2] - 2021-12-18